This fixes search after using VinegarUp (the - key).
Recently it does not always work. Sometimes pressing "-"
opens the parent directory with the cursor on the first
line instead of the current file line.

The actual problem is that the alternative buffer name is
empty after going to the netrw buffer and `expand('#:t')`
returns empty string.

This change fixes it by saving the curren file name before
opening the netrw buffer.

Note that underlying issue is not solved: the alternative
buffer is still empty, so Ctrl-6 to go back still does not work.

Related to https://github.com/tpope/vim-vinegar/issues/136